security: split enterprise authorization policy
