+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -16,3 +16,4 @@ This ledger tracks the enterprise refactor continuation from
 | RISK-REF-009 | Downstream connection lifecycle | Hardened | Both downstream profiles configured pool limits, but each operation constructed and closed a new HTTP client, preventing cross-request connection reuse and leaving shutdown ownership implicit. | Added application-lifespan ownership of reusable dependency-specific HTTP pools, explicit draining/close behavior, and compatibility-preserving injected/standalone paths. Continue reviewing bounded retry posture separately because mutation retries require operation-specific idempotency guarantees. | `tests/unit/test_app_lifecycle.py`; downstream client tests; health tests; `python -m mypy --config-file mypy.ini` |
 | RISK-REF-010 | Runtime configuration and downstream URL safety | Hardened | Core and performance URL resolution duplicated permissive handling; malformed or credential-bearing values failed late, and the required configuration guide was missing. | Added a shared safe URL resolver, negative tests, complete runtime configuration guide, expanded environment example, and security/runbook evidence. | `tests/unit/test_downstream_base_url.py`; `tests/unit/test_configuration_docs.py`; downstream client tests; `make security-audit` |
 | RISK-REF-011 | Enterprise bank runtime enforcement | Hardened | Documented bank mode required authorization and explicit settings, but enabling runtime enforcement did not require that posture and could start permissively. | Made enterprise runtime enforcement fail application construction with bounded issue codes unless the in-process bank posture is complete; kept local defaults unchanged. | `tests/unit/test_enterprise_readiness.py`; deployment-policy and configuration-doc tests; `make security-audit` |
+| RISK-REF-012 | Enterprise authorization module boundary | Hardened | `enterprise_readiness.py` mixed capability-rule parsing, write authorization, runtime validation, redaction, audit emission, payload limits, and middleware assembly, making the security boundary harder to inspect. | Extracted capability-rule parsing and write authorization into `enterprise_authorization.py` while preserving compatibility exports from `enterprise_readiness.py`. | `tests/unit/test_enterprise_readiness.py`; app factory/configuration/deployment-policy tests; `python -m mypy --config-file mypy.ini`; touched radon MI/CC checks |
