+++ src/app/enterprise_readiness.py
@@ -5,6 +5,11 @@ from datetime import datetime, timezone
 from typing import Any, Awaitable, Callable
 
 from fastapi import Request, Response
+from app.enterprise_authorization import (
+    WRITE_METHODS,
+    authorize_write_request,
+    load_capability_rules,
+)
 from app.error_response import error_response
 
 logger = logging.getLogger("enterprise_readiness")
@@ -12,8 +17,6 @@ MiddlewareNext = Callable[[Request], Awaitable[Response]]
 MiddlewareCallable = Callable[[Request, MiddlewareNext], Awaitable[Response]]
 
 _SERVICE_NAME = "lotus-risk"
-_WRITE_METHODS = {"POST", "PUT", "PATCH", "DELETE"}
-_REQUIRED_HEADERS = {"x-actor-id", "x-tenant-id", "x-role", "x-correlation-id"}
 _REDACT_FIELDS = {
     "password",
     "secret",
@@ -119,98 +122,6 @@ def load_feature_flags() -> dict[str, dict[str, dict[str, bool]]]:
     return _load_json_map("ENTERPRISE_FEATURE_FLAGS_JSON")
 
 
-def load_capability_rules() -> dict[str, str]:
-    rules = _load_json_map("ENTERPRISE_CAPABILITY_RULES_JSON")
-    return {key: value for key, value in rules.items() if _valid_capability_rule(key, value)}
-
-
-def _valid_capability_rule(key: Any, value: Any) -> bool:
-    if not isinstance(key, str) or not isinstance(value, str) or not value.strip():
-        return False
-    if key != key.strip() or value != value.strip():
-        return False
-    method, separator, path = key.partition(" ")
-    return bool(separator and method.upper() in _WRITE_METHODS and path.startswith("/"))
-
-
-def _required_capability(method: str, path: str) -> str | None:
-    method = method.upper()
-    matching_rules: list[tuple[int, str]] = []
-    for key, capability in load_capability_rules().items():
-        prefix = f"{method} "
-        rule_path = key[len(prefix) :]
-        if key.upper().startswith(prefix) and _path_matches_rule(path, rule_path):
-            matching_rules.append((len(rule_path.rstrip("/")), capability))
-    if not matching_rules:
-        return None
-    return max(matching_rules, key=lambda item: item[0])[1]
-
-
-def _path_matches_rule(path: str, rule_path: str) -> bool:
-    normalized_rule_path = rule_path.rstrip("/") or "/"
-    return path == normalized_rule_path or path.startswith(f"{normalized_rule_path}/")
-
-
-def _authorization_enforced(method: str) -> bool:
-    return method.upper() in _WRITE_METHODS and _env_enabled("ENTERPRISE_ENFORCE_AUTHZ", "false")
-
-
-def _normalized_headers(headers: dict[str, str]) -> dict[str, str]:
-    return {str(k).lower(): str(v) for k, v in headers.items()}
-
-
-def _missing_required_headers(normalized_headers: dict[str, str]) -> list[str]:
-    return sorted(header for header in _REQUIRED_HEADERS if not normalized_headers.get(header))
-
-
-def _has_service_identity(normalized_headers: dict[str, str]) -> bool:
-    return bool(
-        normalized_headers.get("x-service-identity") or normalized_headers.get("authorization")
-    )
-
-
-def _capability_set(normalized_headers: dict[str, str]) -> set[str]:
-    return {
-        part.strip()
-        for part in normalized_headers.get("x-capabilities", "").split(",")
-        if part.strip()
-    }
-
-
-def _missing_capability_reason(
-    method: str,
-    path: str,
-    normalized_headers: dict[str, str],
-) -> str | None:
-    required_capability = _required_capability(method, path)
-    if required_capability is None:
-        return "missing_capability_rule"
-    if required_capability and required_capability not in _capability_set(normalized_headers):
-        return f"missing_capability:{required_capability}"
-    return None
-
-
-def authorize_write_request(
-    method: str, path: str, headers: dict[str, str]
-) -> tuple[bool, str | None]:
-    if not _authorization_enforced(method):
-        return True, None
-
-    normalized = _normalized_headers(headers)
-    missing = _missing_required_headers(normalized)
-    if missing:
-        return False, f"missing_headers:{','.join(missing)}"
-
-    if not _has_service_identity(normalized):
-        return False, "missing_service_identity"
-
-    missing_capability = _missing_capability_reason(method, path, normalized)
-    if missing_capability:
-        return False, missing_capability
-
-    return True, None
-
-
 def redact_sensitive(value: Any) -> Any:
     if isinstance(value, dict):
         out: dict[str, Any] = {}
@@ -261,7 +172,7 @@ def _content_length(request: Request) -> int:
 
 def _payload_limit_response(request: Request) -> Response | None:
     max_write_payload_bytes = _env_int("ENTERPRISE_MAX_WRITE_PAYLOAD_BYTES", 1_048_576)
-    if request.method not in _WRITE_METHODS:
+    if request.method not in WRITE_METHODS:
         return None
     if _content_length(request) <= max_write_payload_bytes:
         return None
@@ -296,7 +207,7 @@ def _authorization_denied_response(
 
 
 def _emit_write_audit_event(request: Request, response: Response) -> None:
-    if request.method not in _WRITE_METHODS:
+    if request.method not in WRITE_METHODS:
         return
     emit_audit_event(
         action=f"{request.method} {request.url.path}",
@@ -334,3 +245,15 @@ def build_enterprise_audit_middleware() -> MiddlewareCallable:
         return _apply_enterprise_response_headers(response)
 
     return middleware
+
+
+__all__ = [
+    "authorize_write_request",
+    "build_enterprise_audit_middleware",
+    "emit_audit_event",
+    "enterprise_policy_version",
+    "load_capability_rules",
+    "load_feature_flags",
+    "redact_sensitive",
+    "validate_enterprise_runtime_config",
+]
